Be a bit more constant in use of SizeT, and check for error conditions.
git-svn-id: https://ravl.svn.sourceforge.net/svnroot/ravl@7542 7226f50c-d813-0410-99bf-979b288ab488

+++ main/Ravl/Math/LinearAlgebra/General/MatrixDeterminant.cc
@@ -28,7 +28,7 @@ namespace RavlN {
   
   RealT MatrixC::Det() const {
     RavlAssertMsg(Rows() == Cols(),"Can only find the determinant of square matrices.\n");
-    switch(Rows()) {
+    switch(static_cast<size_t>(Rows())) {
     case 0: return 0;
     case 1: return (*this)[0][0];
     case 2: return (*this)[0][0] * (*this)[1][1] - (*this)[1][0] * (*this)[0][1];
